+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8MJYWC2NRA9WW8D20QCH9CT"
+        "01K8N1T114B34A6CEBW4QWKK0Z"
     )
     r"""The unique identifier of the data source"""