+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8MJYWC4G8EM9B4D7NR8XKZY"
+        "01K8N1T116QNYPF8E4CZ2PASZN"
     )
     r"""The unique identifier of the data source"""